pyproject.toml
[project]
name = "meeting-stt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.5",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.52.0",
]
